ref(weekly reports): Rearrange files & functions to be reused
